Updated messages from attribution & distance Mtds
