--- innterpret/attribution/baseline.py
@@ -1,13 +1,0 @@
-from __future__ import absolute_import
-
-# -- IMPORT -- #
-
-# -- ACTIVATION VISUALIZATION METHOD -- #
-class DeepLIFT():
-	def __init__(self):
-		raise NotImplementedError
-
-# -- ACTIVATION VISUALIZATION METHOD -- #
-class SHAPVal():
-	def __init__(self):
-		raise NotImplementedError
