+++ innterpret/attribution/deeplift.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+# -- IMPORT -- #
+
+class DeepLIFT():
+	""">> CLASS:DEEPLIFT: Method Deep LIFT: https://arxiv.org/abs/1704.02685."""
+	def __init__(self):
+		raise NotImplementedError
