--- innterpret/attribution/pattern.py
@@ -1,18 +1,0 @@
-from __future__ import absolute_import
-
-# -- IMPORT -- #
-
-# -- PATTERN -- #
-class Pattern():
-	def __init__(self):
-		raise NotImplementedError
-
-# -- PATTERN NET METHOD -- #
-class PatternNet(Pattern):
-	def __init__(self):
-		raise NotImplementedError
-
-# -- PATTERN ATTRIBUTION METHOD -- #
-class PatternAttribution(Pattern):
-	def __init__(self):
-		raise NotImplementedError
