+++ innterpret/attribution/patternattr.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+# -- IMPORT -- #
+
+class PatternAttribution():
+	""">> CLASS:PATTERNATTRIBUTION: Method Pattern Attribution: http://arxiv.org/abs/1705.05598."""
+	def __init__(self):
+		raise NotImplementedError
