+++ innterpret/attribution/patternet.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+# -- IMPORT -- #
+
+class PatternNet():
+	""">> CLASS:PATTERNNET: Method Pattern Net: http://arxiv.org/abs/1705.05598."""
+	def __init__(self):
+		raise NotImplementedError
