+++ innterpret/attribution/shapvalues.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+# -- IMPORT -- #
+
+class SHAPValues():
+	""">> CLASS:SHAPVALUES: Method Shap Values: https://arxiv.org/abs/1705.07874."""
+	def __init__(self):
+		raise NotImplementedError
