+++ innterpret/distance/tsne.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import
+
+# -- IMPORT -- #
+
+class TSNEDistance():
+    """>> CLASS:TSNEDISTANCE: Method that computes a distribution of the data:
+        https://lvdmaaten.github.io/publications/papers/ JMLR_2008.pdf."""
+    def __init__(self):
+        raise NotImplementedError
