+++ innterpret/utils/bases/metrics.py
@@ -1,4 +1,3 @@
-# -- BASE UTILITIES -- #
 from __future__ import absolute_import
 
 # -- IMPORT -- #
