+++ innterpret/utils/bases/rules.py
@@ -1,108 +1,14 @@
 from __future__ import absolute_import
 
 # -- IMPORT -- #
-from ..utils.tensor import model_remove_softmax, get_model_parameters
-from ..utils.data import load_image, reduce_channels, deprocess_image, visualize_heatmap
 from tensorflow.python.ops import nn_ops, gen_nn_ops
-from keras.models import Model
 import keras.backend as K
-import numpy as np
 import tensorflow as tf
 import copy as cp
+import numpy as np
 
-# -- LRP METHOD -- #
-class LRPModel():
-	def __init__(self,model,verbose=True):
-		self.verbose = verbose
-		print_msg(self.__class__.__name__+' Initialization...',verbose=self.verbose)
-		print_msg('--------------------------',verbose=self.verbose)
-		reluModel = model_remove_softmax(model)
-		_,_,activations,_ = get_model_parameters(model) 
-		self.model = Model(inputs=reluModel.input,outputs=activations)
-		if self.verbose:
-			self.model.summary()
-		self.optionRule = input(set_msg('(ZPlus)-(ZAlpha): '))
-		if self.optionRule == 'ZAlpha':
-			self.alpha = int(input(set_msg('Select a Value for Alpha: ')))
-			assert self.alpha >= 1
-		self.numLayers = len(self.model.layers)
-		print_msg('========== DONE ==========\n',verbose=verbose)
-
-	# >> DEFINE_RULES: goes through the model to save the specified rules
-	def define_rules(self,imgData,oneHot=False):
-		print_msg('Define Rules...',verbose=verbose)
-		print_msg('--------------------------',verbose=verbose)       
-		layerRules = []; self.outputR = self.model.predict(imgData)
-		if oneHot:
-			neuron = np.argmax(self.outputR[-1])
-			maxAct = self.outputR[-1][:,neuron]
-			self.outputR[-1] = np.zeros(self.outputR[-1].shape,dtype='float32')
-			self.outputR[-1][:,neuron] = maxAct
-		for currLayer,k in zip(reversed(self.model.layers),range(self.numLayers-2,-1,-1)):
-			nextLayer = currLayer._inbound_nodes[0].inbound_layers[0]
-			activation = self.outputR[k-1] if (k-1!=-1) else imgData
-			if self.optionRule == 'ZPlus':
-				layerRules.append(ZPlus(currLayer,activation))
-			elif self.optionRule == 'ZAlpha':
-				layerRules.append(ZAlpha(currLayer,activation,self.alpha))
-			else:
-				assert False, 'This Rule Option is not supported'
-			print_msg('<><><><><>',verbose=verbose)
-			print_msg('Weights From: ['+currLayer.name+']',verbose=verbose)
-			print_msg('Activations From: ['+nextLayer.name+']',verbose=verbose)
-		self.rules = layerRules
-		print_msg('========== DONE ==========\n',verbose=verbose)
-
-	# >> RUN_RULES: computes the relevance tensor for all the model layers.
-	def run_rules(self):
-		print_msg('Run Rules...',verbose=verbose)
-		print_msg('--------------------------',verbose=verbose)
-		R = {};
-		R[self.rules[0].name] = K.identity(self.outputR[-1])
-		self.print_rule_information(R[self.rules[0].name],verbose=verbose)
-		for k in range(len(self.rules)):
-			if k != len(self.rules)-1:
-				R[self.rules[k+1].name] = self.rules[k].run(R[self.rules[k].name],ignoreBias=False)
-				self.print_rule_information(R[self.rules[k+1].name],verbose=verbose)
-			else:
-				R['input'] = self.rules[k].run(R[self.rules[k].name],ignoreBias=False)
-				self.print_rule_information(R['input'],verbose=verbose)
-		print_msg('========== DONE ==========\n',verbose=verbose)
-
-	def print_rule_information(R):
-		""">> PRINT_RULE_INFORMATION: prints out the tensor shape of the rules runned."""
-		if verbose:
-			shape = print_tensor_shape(R)
-			print_msg('<><><><><>')
-			print_msg('Rule R[input] Correctly runned.')
-			print_msg('Tensor Output Shape: '+shape)
-
-	# >> VISUALIZE: returns a graph with the results.
-	def visualize(self,savePath,cmap='plasma',option='sum'):
-		print_msg('Visualize '+self.__class__.__name__+' Result...')
-		print_msg('--------------------------')
-		oneDim = reduce_channels(self.relevance.copy(),option=option)
-		heatMap = deprocess_image(oneDim.copy())
-		visualize_heatmap(self.rawData,heatMap[0],self.__class__.__name__,cmap,savePath)
-		print_msg('========== DONE ==========\n',verbose=verbose)
-
-	# >> EXECUTE: returns the result of the LRP method
-	def execute(self,fileName,sess,layerName=None,oneHot=False):
-		print_msg(self.__class__.__name__+' Analyzing')
-		print_msg('--------------------------')
-		self.rawData = load_image(fileName,preprocess=False)
-		imgData = load_image(fileName)
-		self.define_rules(imgData,oneHot=oneHot,verbose=verbose)
-		self.run_rules(verbose=verbose)
-		if layerName is None:
-			self.relevance = sess.run(self.R['input'])
-		else:
-			self.relevance = sess.run(self.R[layerName])
-		print_msg('========== DONE ==========\n')
-		return self.relevance
-
-# -- Z+ RULE -- #
 class ZPlus(object):
+	""">> CLASS:ZPLUS: Rule to compute the relevance propagation."""
 	def __init__(self,currLayer,nextAct):
 		self.layer = currLayer
 		self.name = currLayer.name
@@ -111,7 +17,8 @@ class ZPlus(object):
 		self.maxValue = K.epsilon()
 		self.minValue = -K.epsilon()
 		
-	def run(self,R,ignoreBias=True):
+	def run(self,R,ignoreBias=False):
+		""">> RUN: run the relevance propagation for the current layer."""
 		if self.type == 'Dense':
 			return self.run_dense(R,ignoreBias)
 		elif self.type == 'MaxPooling2D':
@@ -123,7 +30,8 @@ class ZPlus(object):
 		else:
 			raise NotImplementedError
 			
-	def run_dense(self,R,ignoreBias=True):
+	def run_dense(self,R,ignoreBias=False):
+		""">> RUN_DENSE: Run the relevance propagation for dense layers."""
 		weights = self.layer.get_weights()
 		self.W = K.maximum(weights[0],0.)
 		self.B = K.maximum(weights[1],0.)
@@ -135,11 +43,13 @@ class ZPlus(object):
 		return K.clip(self.act*C,self.minValue,self.maxValue)
 	
 	def run_flatten(self,R):
+		""">> RUN_FLATTEN: Run the relevance propagation for flatten layers."""
 		shape = self.act.get_shape().as_list()
 		shape[0] = -1
 		return K.reshape(R, shape)
 	
 	def run_pool(self,R):
+		""">> RUN_POOL: Run the relevance propagation for pool layers."""
 		poolSize = (1,self.layer.pool_size[0],self.layer.pool_size[1],1)
 		strdSize = (1,self.layer.strides[0],self.layer.strides[1],1)
 		pooled = tf.nn.max_pool(self.act, 
@@ -154,6 +64,7 @@ class ZPlus(object):
 		return K.clip(self.act*C,self.minValue,self.maxValue)
 	
 	def run_conv(self,R,ignoreBias):
+		""">> RUN_CONV: Run the relevance propagation for conv layers."""
 		strdSize = (1,self.layer.strides[0],self.layer.strides[1],1)
 		weights = self.layer.get_weights()
 		self.W = K.maximum(weights[0],0.)
@@ -170,8 +81,8 @@ class ZPlus(object):
 										  S,strdSize,self.layer.padding.upper())
 		return K.clip(self.act*C,self.minValue,self.maxValue)
 
-# -- ZAlpha RULE -- #
 class ZAlpha(object):
+	""">> CLASS:ZALPHA: Rule to compute the relevance propagation."""
 	def __init__(self,currLayer,nextAct,alpha):
 		self.alpha = alpha
 		self.beta = 1-alpha
@@ -182,7 +93,8 @@ class ZAlpha(object):
 		self.maxValue = K.epsilon()
 		self.minValue = -K.epsilon()
 		
-	def run(self,R,ignoreBias=True):
+	def run(self,R,ignoreBias=False):
+		""">> RUN: run the relevance propagation for the current layer."""
 		if self.type == 'Dense':
 			return self.run_dense(R,ignoreBias)
 		elif self.type == 'MaxPooling2D':
@@ -194,7 +106,8 @@ class ZAlpha(object):
 		else:
 			raise NotImplementedError
 		
-	def run_dense(self,R,ignoreBias=True):
+	def run_dense(self,R,ignoreBias=False):
+		""">> RUN_DENSE: Run the relevance propagation for dense layers."""
 		weights = self.layer.get_weights()
 		self.maxW = K.maximum(weights[0],0.); self.maxB = K.maximum(weights[1],0.)
 		self.minW = K.minimum(weights[0],0.); self.minB = K.minimum(weights[1],0.)   
@@ -207,11 +120,13 @@ class ZAlpha(object):
 		return K.clip(Rn,self.minValue,self.maxValue)
 		
 	def run_flatten(self,R):
+		""">> RUN_FLATTEN: Run the relevance propagation for flatten layers."""
 		shape = self.act.get_shape().as_list()
 		shape[0] = -1
 		return K.reshape(R, shape)
 	
 	def run_pool(self,R):
+		""">> RUN_POOL: Run the relevance propagation for pool layers."""
 		poolSize = (1,self.layer.pool_size[0],self.layer.pool_size[1],1)
 		strdSize = (1,self.layer.strides[0],self.layer.strides[1],1)
 		pooled = tf.nn.max_pool(self.act, 
@@ -230,6 +145,7 @@ class ZAlpha(object):
 		return K.clip(Rn,self.minValue,self.maxValue)
 	
 	def run_conv(self,R,ignoreBias=True):
+		""">> RUN_CONV: Run the relevance propagation for conv layers."""
 		strdSize = (1,self.layer.strides[0],self.layer.strides[1],1)
 		weights = self.layer.get_weights()
 		self.maxW = K.maximum(weights[0],0.); self.maxB = K.maximum(weights[1],0.)
